Use LK api instead of webhook to track sessions
Since the webhook does not return all the necessary session data, we need to additionally request it via API, but it may become available there with a significant interval after the webhook. Therefore, instead, we request via API a list of all sessions that were created since yesterday, including unfinished ones, and update the information in billing.

Signed-off-by: Anton Alexeyev <alexeyev.anton@gmail.com>